Making color service improvements
