+++ src/services/color/color-sets/keppel-color-set.ts
@@ -1,0 +1,36 @@
+export const KEPPEL_COLOR_SET = {
+    'primary': '#00a7a2',
+    'accent': '#7b63a3',
+    'secondary': '#fff',
+    'alternate1': '#3baa43',
+    'alternate2': '#025662',
+    'alternate3': '#b08f5c',
+    'vibrant1': '#00cceb',
+    'vibrant2': '#ff9048',
+    'grey1': '#2a2a2a',
+    'grey2': '#333',
+    'grey3': '#666',
+    'grey4': '#999',
+    'grey5': '#ccc',
+    'grey6': '#eee',
+    'grey7': '#f5f5f5',
+    'grey8': '#fafafa',
+    'chart1': '#00a7a2',
+    'chart2': '#7b63a3',
+    'chart3': '#3baa43',
+    'chart4': '#025662',
+    'chart5': '#b08f5c',
+    'chart6': '#ccc',
+    'ok': '#3baa43',
+    'warning': '#ff9048',
+    'critical': '#ff454f',
+    'partition1': '#635387',
+    'partition9': '#4a4066',
+    'partition10': '#308935',
+    'partition11': '#023e42',
+    'partition12': '#91744d',
+    'partition13': '#999',
+    'partition14': '#294266',
+    'social-chart-node': '#00cceb',
+    'social-chart-edge': '#00cceb'
+};
