+++ src/services/color/color-sets/micro-focus-color-set.ts
@@ -1,0 +1,79 @@
+export const MICRO_FOCUS_COLOR_SET = {
+    'cerulean': '#1668c1',
+    'aqua': '#29ceff',
+    'aquamarine': '#2fd6c3',
+    'fuchsia': '#c6179d',
+    'indigo': '#7425ad',
+    'dark-blue': '#231ca5',
+    'white': '#ffffff',
+    'slightly-gray': '#f5f7f8',
+    'bright-gray': '#f1f2f3',
+    'gray': '#dcdedf',
+    'silver': '#bdbec0',
+    'dim-gray': '#656668',
+    'dark-gray': '#323435',
+    'black': '#000000',
+    'crimson-negative': '#e5004c',
+    'apricot': '#f48b34',
+    'yellow': '#fcdb1f',
+    'green-positive': '#1aac60',
+    'ultramarine': '#3939c6',
+    'skyblue': '#00abf3',
+    'pale-aqua': '#43e4ff',
+    'pale-green': '#1ffbba',
+    'lime': '#75da4d',
+    'orange': '#ffce00',
+    'magenta': '#eb23c2',
+    'pale-purple': '#ba47e2',
+    'dark-ultramarine': '#271782',
+    'steelblue': '#014272',
+    'arctic-blue': '#0b8eac',
+    'emerald': '#00a989',
+    'olive': '#5bba36',
+    'goldenrod': '#ffb000',
+    'purple': '#9b1e83',
+    'pale-eggplant': '#5216ac',
+    'red': '#ff454f',
+    'pale-amber': '#ffb24d',
+    'pale-lemon': '#fde159',
+    'pale-emerald': '#33c180',
+    'plum': '#b21646',
+    'copper': '#e57828',
+    'amber': '#ffc002',
+    'leaf-green': '#118c4f',
+    'forest-green': '#00645a',
+    'primary': '#0073e7',
+    'accent': '#7425ad',
+    'secondary': '#ffffff',
+    'alternate1': '#29ceff',
+    'alternate2': '#2fd6c3',
+    'alternate3': '#c6179d',
+    'vibrant1': '#43e4ff',
+    'vibrant2': '#ffce00',
+    'grey1': '#000000',
+    'grey2': '#323435',
+    'grey3': '#656668',
+    'grey4': '#bdbec0',
+    'grey5': '#dcdedf',
+    'grey6': '#f1f2f3',
+    'grey7': '#f5f7f8',
+    'grey8': '#ffffff',
+    'chart1': '#3939c6',
+    'chart2': '#00abf3',
+    'chart3': '#75da4d',
+    'chart4': '#ffce00',
+    'chart5': '#eb23c2',
+    'chart6': '#ba47e2',
+    'ok': '#1aac60',
+    'warning': '#f48b34',
+    'critical': 'e5004c',
+    'partition1': '#7425ad',
+    'partition9': '#5216ac',
+    'partition10': '#5bba36',
+    'partition11': '#014272',
+    'partition12': '#ffb000',
+    'partition13': '#bdbec0',
+    'partition14': '#271782',
+    'social-chart-node': '#ff00ff',
+    'social-chart-edge': '#ff00ff'
+};
