+++ src/services/color/index.ts
@@ -1,2 +1,4 @@
+export * from './color-sets/index';
 export * from './color.module';
 export * from './color.service';
+export * from './theme-color';
